correcion clase 1
